fix: 🐛 Improving Logger
