--- src/controllers/check-email.controller.ts
@@ -1,11 +1,0 @@
-import { Controller, Post, Request } from '@nestjs/common';
-import { CheckEmailService } from '../services/check-email.service';
-@Controller()
-export class AuthController {
-  constructor(private readonly checkEmailService: CheckEmailService) {}
-
-  @Post('/auth/check-email')
-  getHello(): string {
-    return this.checkEmailService.getHello();
-  }
-}
